Implement Focus API, fixes #6
The Focus API has been implemented, allowing a component
to pass on mouse & keyboard events to a sole component,
rather than propogating normally.

In order for this to work, it required refactoring the mouse
enter & leave events to propogate in the same fashion that the
rest of the events do.