UIMultilineTextInput: Fix space overflowing the right boundary
Previously, I intentionally made a change which allowed a space to be
included in visual lines such that they overflowed the right side of the
component. I did this to avoid rendering a space at the beginning of a
line (if there is only one space; of course, if there are many spaces in
a row, they will be rendered at the beginning).

The change was good, but had one negative side effect: the cursor could
overflow the right boundary. To fix this, I prevented the cursor from
being drawn to the right of the boundary. However, this resulted in two
cursor positions that were very close to each other.

This fix changes it so that if a space *would have* overflowed the
boundary, the entire previous word is wrapped to the new line.