Inspector: Implement component hit testing & hidden checking
Now, the Inspector element has a button on the titlebar that, when
clicked, will allow the user to move their mouse across the screen,
drawing the outline box over the currently hovered component.

Additionally, this patch changes the Inspector's tree view arrows
to be plus/minus boxes, which is more commonly seen in tree views.

Finally, this patch also adds text to an InspectorNode to indicate
that a component is hidden.