SVGComponent: Parse and apply viewBox property
Just the width and height part of it though, as x and y don't really
mean anything outside of the DOM.

This makes the actual "width" and "height" property of an SVG
essentially useless, as in Elementa we specify width and height via
constraints -- it should be the same way for SVGs.

Previously, the size of the rendered svg depended on those properties,
so to change the size you would have to make sure those properties were
set correctly; very unintuitive!