ConstraintResolver: Only resolve on StackOverflowError
This reduces the amount of times the whole constraint tree is validated,
and makes performance much less of a concern (as it will only happen on
error). Additionally, print log message (and in the future, a chat
message) if this happen when the elementa.dev environment variable is
not set.