UIComponent/WindowScreen: Add an afterInitialization function
This functions runs immediately before the first draw invocation. This
is useful for running any code that needs to be run on setup, but would
crash if put in the init block (such as any animation API interactions),
as by the time it is called the component hierarchy has been fully
setup.