UIWrappedText: Add truncation; Text Components: Consider text scale
UIWrappedText now accepts a trimText option which, when set to true,
will truncate the text to be strictly inside the bounds of its width and
height constraints. If the text is truncated, an ellipsis is added to
the end.

Additionally, many places within the various text component classes did
not take text scale into account (mostly for {String,Char}.width()
calls). This has been fixed.