MarkdownComponent: Begin a complete rewrite
With the need to center markdown in ModCore, it was apparent that the
first iteration of this component wasn't all that great. I am rewriting
it from the ground up with the intent of readability and extendability,
the first step of which is using a third-party library,
commonmark-java, for markdown parsing. Markdown parsing is extremely
difficult, and it's not worth it to figure it out myself when I can
just use a small library for it.