MarkdownComponent: Begin selection support
Selection is a two-fold process. This commit introduces support for the
first step.

The first step in selecting text is being able to turn a mouse position
into a textual position inside of a drawable. This is done by the
DrawableCursor class. Each Drawable type will have its own
implementation of the DrawableCursor class which knows how to navigate
that specific component. The MarkdownCursor class is responsible for
managing cursors for the entire MarkdownComponent (e.g. it will pick
the correct cursor class given an arbitrary mouse position inside of
the MarkdownComponent).

The second step in the selection process is, given two cursors,
determine the string "between" them which should be copied to the
user's clipboard. This will be handled in a similar way to the cursors;
each drawable will have a coresponding DrawableSelection class, and
the MarkdownSelection class will coordinate multiple selection classes.
This may change in the future.