MarkdownComponent: Begin working on links
This works pretty well, but if the same link is wrapped across a
newline, both sides of the link won't be hovered. Additionally, it
would be nice if we showed a popup when clicking a link instead of
just opening it in the browser. It should be conditional though;
we wouldn't want it opening on ModCore's welcome screen, for example