UIImage: Fix cache being corrupted due to incorrect fetch
By using `compute` the cached entry was being replaced by newly created ones,
even though it was only meant to be retrieved and not replaced.
This caused assets for which multiple (2 or more) UIImages were requested before
the asset was loaded to never fully load because any image after the second one
used the wrong cached instance.