Constraints: Always lay out Elements aligned with real pixels
Most of our rendering does not supported sub-pixel positions and will give
inconsistent and bad results. Unfortunately it was really easy to accidentally
render a component at such a position, simply adding a RelativeContraint, a
CenterConstraint, a 0.25 pixel constraint, or similar is enough.

This commit prevents such situations (or at least makes them much more difficult
to induce on accident) by rounding the result of any X/Y/Width/Height
constraints to full pixels (real pixels, not the scaled MC pixels, so animations
and centering should still look reasonably good, even at higher GUI scales).
The same guard is added to the VanillaFontRenderer because it is common for
consumer components to call that method with X/Y offset by half of the
width/height of the container to center the text, which may very well result in
a sub-pixel position.

This change is almost non-breaking except in cases where a consumer component or
constraint relies on the sub-pixel layout behavior for correctness (such as a
grid layout which does not take pixels into account).

The CenterConstraint gets special treatment because it is so common and
otherwise the direction it rounds towards will be semi-random depending on the
position (cause of floating point math), even when the heights are the same,
which is not ideal and noticeable when re-sizing the window or (presumably) with
slow animations.