UIShape: Add convex-only warning
As per GL spec, GL_POLYGON only supports convex polygons. But it may still render
fine with concave ones which makes it difficult to be aware of this
restriction (our own example misuses it).
To make this more obvious, this commit deprecates the component until it
implements its own triangulation which supports concave shapes.