Inspector: Fix node index being incorrect
In particular when the new component is inserted between two existing components
or when it has the Inspector component (which we hide) as a sibling somewhere
before it.
This could result in the wrong node or an out of bounds node being removed,
resulting in incorrect behavior or an immediate crash.