ConstraintResolutionGui: Display the component class and name
This is the way the Inspector does it, and I think it's much better.
The hex code was to uniquely identifier the component in the list view,
but if the user is having trouble determining the components from the
list view, they should either look at the tree view or use delegation
to set the component names.