UITextInput: Fix cursor being visible while inactive
Happened when there was an active selection when the input lost focus. In this
case the cursor is hidden, so the `unhide` call in `setCursorPos` would trigger
the unhide animations (setting color to blinking), whereas if there was no
selection, then the cursor would be set to invisible in `setActive` but would
not be hidden and as such the `unhide` call had no effect.