UIText: Fix impure font width state mapping
Having an impure map function invalidates the whole point of states because they
will not update when the impure parts change.

Someone tried to work around it by rebinding the state on the next render but
that does nothing after the first render and just obfuscates the original issue.
The proper way to fix this is to not have impure inputs in the first place.