MarkdownComponent: Remove reparse on document re-layout
This fixes an issue that has been in the MarkdownComponent for a long
time, which is that whenever the MarkdownComponent re-laid itself out,
it also reparsed the text. This was because the ParagraphDrawable
mutated its drawable list, meaning that calling `layout()` without
reparsing would cause the ParagraphDrawable to eventually have a ton
of really tiny text components (since they would be split down and
never joined).

The quick and dirty fix to this was to simply reparse the entire
document, which would regenerate the ParagraphDrawable with brand-new
TextComponents. Simply not mutating the drawable list in
ParagraphDrawable was not a hard change to make, so I'm not sure why
I didn't do this from the beginning.

Not only does this make the code nicer _and_ increase performance, it
also fixes a crash. When an inline image was loaded in the markdown,
the actual image loading would happen in a CompletableFuture. But when
the image loaded and started to be drawn, it would cause a relayout
(due to the drastic height change), causing the image component to be
thrown away. The thrown-away image would attempt to access its parent,
which no longer existed, and would crash. This no longer happens!