ParagraphDrawable: Account for HardBreakDrawables
Previously this was just a TODO() if one occurred in the Drawables
list because I wasn't sure if it was possible, but it definitely is.
The easiest repro is a backslash before newlines, which apparently
is a hard-break (TIL!). Another repro is multiple inline images on
their own lines, one after each other.