SiblingConstraint: Fix incorrect padding when first child
When a SiblingConstraint is applied to the first child of a container, then the
child is positioned at in the very top left (or bottom right if opposite) of the
container, i.e. the padding is zero. This does not apply if the
SiblingConstraint is bound to a specific target; in that case the padding always
applies.

This error went mostly unnoticed because the ChildBasedSizeConstraint (which is
the primary consumer of these padding methods) is broken as well in that it
ignores the padding of the last child, thereby canceling out this issue most of
the time. That issue will be fixed in the following commit, thereby restoring
the balance again for the vast majority of users.
I am not considering this change to be breaking because the previous
behavior (if you ran into it) was unambiguously incorrect and no one should have
relied on it in the first place.