ChildBasedSizeConstraint: Fix padding of last child being ignored
This error went mostly unnoticed because the SiblingConstraint (which is the
primary provider of these padding methods) is broken as well in that it returns
surplus padding for the first child, thereby canceling out this issue most of
the time. That issue was fixed in the previous commit, therefore the balance is
restored again with this commit for the vast majority of users.
I am not considering this change to be breaking because the previous
behavior (if you ran into it) was unambiguously incorrect and no one should have
relied on it in the first place.