Inspector: Fix incorrect Window used
For the constraints, the inspector should use the window which it itself is a
part of, not the window of the thing it is inspecting.

For the click interceptor, that's debatable because the current behavior does
not work with embedded windows whereas the new behavior does not work when we
support detaching the inspector into its own window. For the time being we care
only about the former, so we'll go with that.