ConstraintResolutionGui: Allow inspecting of broken window
This commit adds a button which shows the broken window in a frozen
form (rendering only) with a fully interactive Inspector window.
The constraint debugger system is used to resolve cyclic constraints to an
arbitrary result of 10 so the whole thing can be viewed and inspected without
exploding.

It also adds a button to each entry in the cyclic stack which will switch to the
broken window and focus the corresponding element in the inspector. The graph
view entries are also made clickable with the same behavior.