UIText: Fix side-effect in `State.map` call
All `State.map` calls should be free of side-effects. If you need to update
some other state (as is the case here) whenever one of its inputs changes, that
other state should instead be derived from its input, so it will automatically
refresh when any of its inputs are changed.