ScrollComponent: Fix scroll bar with padding
Taking the absolute value of the offset gives incorrect results when the scroll
component has an inner padding. In such cases, the offset will be in range e.g.
`[-1000, 15]`, so if scrolled all the way to the top, the value would be 15 and
the scroll bar would not be placed at 0% where it should be.
Instead we must subtract the offset from the padding (e.g. `15 - 15 = 0`).