MarkdownComponent: Stop abusing afterInitialization method
There's no need for MarkdownComponent to override it, and then having to
manually call it because it can't call super.draw because that's too later for
when it wishes to draw.

Also happens to fix effect initialization (which happened in the original
afterInitialization), the effect before draw callback (which was forgotten
about), incorrect timing of the effect before children draw callback (which was
called before the component is draw) and anything which requires `super.draw` to
function correctly (e.g. debug outlines).