MarkdownComponent: Layout in animationFrame instead of draw
Cause that's the place where layout is supposed to take place. One reason for
why this is the better place is that `draw` isn't called until the component is
drawn, and elementa doesn't draw components if they're not visible, therefore
markdown components where pretty much unusable with scrollbars cause their size
would be all wrong.