ExampleMod: Convert to Java
Because that compiles faster and this class will be the only major
version-specific class we'll have remaining once we split versions-specific from
non-specific code.
Originally I did this because I had hoped to be able to get rid of Kotlin
in the version-specific project tree altogether but that turned out to not be
possible, so the actual gains from this change may not be particularly
measurable but it's implemented, so might as well commit it.