ArrowComponent: Utilized cached UIImage instead of SVG
The ArrowComponent previously read and parsed the open and close SVG each time an instance is created. This severely degraded the performance of the inspector whenever the layout changed, and did not work on modern versions. The component now utilizes a cached UIImage for both better performance and increased compatibility.