UIImage: Address race condition in texture loading.
Previously, UIImage would set its `texture` field as soon as it's available from the future callback. However, the texture was not uploaded until the next frame via `Window.enqueueRenderOperation`. As a result, calls to `supply` could apply the texture before the texture was uploaded. This has been fixed by not setting the field until the texture is uploaded.