UIComponent: Lazy init component name (#84)
Except when debugging, the component name does not need to be resolved and the performance hit of eagerly resolving the name via the java class is non-negligible. To improve this behavior, the value is defaulted to some marker value and resolved via java class only when needed. If the component name is defined via the component delegated properly, then the call to the java class is eliminated entirely.