Window: Fix live-lock when `animationFrame` is slow
If the window is sufficiently complex, it's possible for the average
`animationFrame` to take so long we'll start falling behind with no way to ever
catch up. And the amount of frames we're behind will quickly grow to the point
where we'll be spending five seconds in `animationFrame` before we can get a
real frame on the screen.
Or even worse with the old while-loop, we'll be stuck indefinitely because the
five seconds check only happens outside the loop and we never get to exit the
loop if we never catch up, effectively locking up the game.

To prevent that, we limit the `animationFrame` calls we make per real frame such
that we'll still be able to render approximately 30 real frames per second at
the cost of animations slowing down.