OverridableAlphaColorConstraint: Recalculate original constraint's color
We must forward the constraint's animation frame to the original's. We
want colours to be recalculated when needed and this prevents that
during the `RecursiveFadeEffect`'s lifecycle.

The `originalAlpha` member was also removed. It didn't appear like there
was any reason for caching the alpha, and a use-case where someone might
depend on that is quite niche (neither I nor Johni could think of one),
and its removal is required for this change.