Window: Fix incorrect `animationFrames` computation
We expect the `animationFrames` variable to contain the amount of frames to run
but what we used to compute does not even have the correct units:
`ms / (frames / sec)`

This wasn't immediately obvious because the further we drift from the target,
the closer this value actually becomes to the correct value, so after just a few
seconds it was no longer noticeable.

This commit fixes that. The new computation is
`ms * (frames / sec) / (ms / sec) = ms * frames / sec * sec / ms` which all
nicely cancels out leaving us with only `frames` as expected.