UIComponent: Call `afterDraw` on effects in reverse-order
Prior to this, if you had two effects (e.g. ScissorEffect) of the same
type on the same component, they would clash due to `afterDraw` being
called in the same order as `beforeDraw`. Calling `afterDraw` on these
effects in reverse solves the problem.

This also introduces a new Elementa version, V3 - which deprecates V2.

GitHub: #118