StateV2: Re-implement `toV2` with more indirection
This way any listeners registered on the returned v2 State are weak again (as
would be expected from any v2 State), and only the v2 State itself stays
registered on the v1 state indefinitely.

Also allows us to move away from `onSetValue` as a primitive because we no
longer pass that callback to the v1 directly.

Source-Commit: 43a0362a909429224b0ca267f1be3cb33695a1c9