StateV2: Fix effect staying alive when unregistered from func
When `cleanup` was called for an effect from the `func` of that effect, then
after the `func` returned, we'd immediately re-subscribe it to its dependencies,
thereby making the `cleanup` ineffective and keeping the effect alive for longer
than it is supposed to be.

This commit fixes the issue by introducing `NodeState.Dead` which is set when the
node gets cleaned up. We can then skip re-subscribing when the node is in this
state.

Source-Commit: 840aaf6988d47a0b50401b376c5ea5acc4b98097