LayoutDSL: Add layoutAsBox/Row/Column entrypoints
Unlike the raw `layout`, these will provide the expected defaults for children
positioning.
It must however be noted that they do not change the size of the target
component, because constraints of such plain components have often already been
set up beforehand, often with more complex expressions which we do not want to
overwrite.

Source-Commit: cb90448da5cf8d051ae5dc4c7b7e93d6315960b1