LayoutDSL: Fix Arrangement not being reusable
Prior to this commit it was not possible to use one `Arrangement` instance for
multiple e.g. rows. This was unexpected because it looks from the outside to be
reusable, doubly so because `Modifier` is reusable.

This commit fixes the issue by renaming `Arrangement` to `ArrangementInstance`
and putting a factory style interface in its place instead.

Source-Commit: 8a7df4692856360641e93d4ad0ba2184489e2b53