LayoutScope: Give uncached for-each child scopes their own `stateScope`
Prior to this, these scopes would use the `component` of the LayoutScope
that the `forEach` is called on as their `stateScope`.

This is incorrect though, as if something in the `block` happens to use
said `stateScope`, and the entry gets removed from the `ListState<T>`,
the `stateScope` will still be alive.

With these changes, the `stateScope` for these entries is now disposed
appropriately.

`stateScope` was also used as an escape hatch for accessing the
`container` of a `LayoutScope`; this is considered bad practice because
it is similarly easy to use a component one shouldn't be messing with.
However sometimes access to the component is required, so another
escape hatch has been added to serve these cases.

Source-Commit: 270363df989772e99667359e9eaedb15b916149e