Build: Use Standalone UniversalCraft for example
This has the advantage that it's much faster to boot than full MC and
that one is no longer required to set up a Minecraft development
enviornment for multiple versions to build Elementa.

Also replaces the example image url because imgur was redirecting to its
html page instead of returning the image.