StateV2: Fix `ReferenceHolder.Weak` usage in `toV1` and `await`
Prior to the graph-based State implementation, keeping a reference to
the listener callback object was sufficient to prevent the underlying
subscription from being garbage collected.
With the new implementation however, this is no longer the case and one
needs to instead hold on to the returned `unregister` function to keep
the `effect` alive.

Our `State.toV1` and `State.await` implementations were still relying on
this old behavior.

For `await` the issue is less significant because we already held on to
the unregister function from the listener function, so it will function
properly but may keep the referenced state alive for longer than
necessary (depending on how the coroutine holds on to the locals after
its completion).

For `toV1` however, if used with `ReferenceHolder.Weak`, this will
result in updates breaking completely after
garbage collection.

This commit fixes both of these by switching to `effect`.
It also removes the documentation of the long deprecated `onSetValue`
method because it still describes the old behavior which is incorrect
now.

Source-Commit: eff73d6aca8289feb1f33bf90c44e54ea81dc70e