StateV2: Fix toV2 on weak V1 state
The implementation of `toV2` did not consider weak v1 State
implementations (because Elementa itself does not provide any), and as
such was relying purely on the strong reference from the callback
registered on the input v1 State to keep its references alive.
If that v1 State however is weak, then it won't be sufficient to keep
everything alive. Instead we need to keep our internals alive as long as
the State we return may still be used, i.e. as long as there is some
reference to it. The fix in this commit therefore is to return a State
which includes a strong references to the internal ReferenceHolder.

Source-Commit: 5b18839e5482c0503ae4722ba4d6883c11495c86