StateV2: Fix `await` cancellation
Previous implementation would simply drop the unregister function, which
works only under the incorrect assumption that coroutines will drop all
resume calls after cancellation.
However, coroutines are smarter than that, and while they will ignore
the first resume call if the job was already cancelled, it will still
verify the resume-must-only-be-called-once contract even in those cases
and yell at you if you call it a second time.

This commit fixes that by not dropping unregister in general but a
separate reference specifically for this task. The effect can then still
unregister itself after it was called for the first time and won't be
called a second time.

Source-Commit: 5fedf3858046e0ca7e5d5607e62a20504dea56f4