UIComponent: Deprecate `animationFrame`
See the kdocs on `ElementaVersion.V8` for details and motivation.

Aside from the changes behind the ElementaVersion flag, this commit also
changes the behavior of constraints evaluated on components which do not
have a window (because we therefore cannot determine the
ElementaVersion) to never cache its value.
This is necessary because if we were to cache its value in that
situation, we wouldn't be able to get it registered in the Window's
`cachedConstraints` list, and so it'd never be invalidated again on V8+.
The chance of someone relying on the constraint caching behaivor in such
a specific scenario are sufficiently tiny that it seems safe to make
this change.