animateTransitions: Fix unncessary extra initial transition
The `AnimationDriver` implementation currently emits an unnecessary
event which does not change the value into its `animationEventList` on
initialization.
This commit remove that unnecessary event by using `onChange` instead of
`effect`; we already query the current value of the state two lines up,
so should be safe to ignore it here.

Source-Commit: 17be5cc34aa0a0ea1e9864a653314e0f3b903db7