UIComponent: Rewrite floating API
The old API is difficult to work with because it will continue to render
floating components even if they've been removed from the component
tree. Additionally it can CME if components are removed from the
floating list during rendering, further complicating the workarounds
required.

This new API fixes the issue by tracking when components are
removed/added from/to the tree and updating its internal floating list
accordingly.
It also allows setting the floating state at any time, even before the
component has a parent, another thing the old API did not support.

The order in which floating components appear also differs in the new
API. While the old API showed floating components in the order in which
they were set to be floating, this often isn't all too useful when the
order in which components are added/removed to/from the tree is not
particularly well defined. As such, the new API chooses to instead order
floating components in exactly the same way as they appear in the
component tree (pre-order tree traversal, i.e. first parent, then
children). This results in consistent ordering and is generally the
order you want for nested floating components to behave in a useful way.

This has been implemented as a new, completely separate API instead of
an ElementaVersion primarily to easy migration (the new API can be used
even with Windows still on older ElementaVersions; both APIs can be used
at the same time) but also because there isn't anything reasonable the
old-API methods in `Window` could do in the new version, they really
should have been internal to begin with.

GitHub: #154