Inspector: Show source file of components and jump there in IntelliJ
This commit tries to track the source code location of components (not
where its class is defined but where that particular instance was
created) and shows that info in the inspector. Should the automatically
determined location be unhelpful (e.g. because it's in a factory
function rather than where the factory function is called), then it also
allows scrolling while holding the Shift key to go up and down the call
stack.

If IntelliJ is running, it will also automatically jump to that source
location every time a component is selected in the inspector.