UIComponent: Optimize mouseMove and dragMouse methods
Previously these would both just traverse the entire tree and evaluate
the constraints of all components.
In Essential, only components which can be dragged care about drag
events, so most components don't, and since we primarily use the
`hoveredState` from unstable Elementa, virtually no components need the
mouseEnter/Leave listeners (which is what mouseMove is for).

This commit thusly optimizes the two methods to only traverse the narrow
subtree which has components that might be interested.
For simplicity, this is a best-effort optimization (e.g. it may still
visit components which have had such listeners at some point but no
longer do now); it should however be completely correct in that it will
behave identical to how it behaved prior to this commit, just consume
less CPU where easily possible.

Only exception are the `lastDraggedMouseX/Y` properties which are
impossible to accurately emulate without traversing every component.
They should really have been private and only in Window to begin with...
This commit compromises by accepting that their behavior may be
different now if set manually or if dragMouse is called manually, but
still preserves the rough meaning if it's merely read from in e.g.
`mouseRelease`. The exact behavior of these properties was sufficiently
unexpected that hopefully no one will have tried to rely on their exact
behavior to begin with.

Note that the Flags class and tracking introduced in this commit also
support `Effect`s, despite neither of the two methods being available in
`Effect`s. This is because we'll also use the same mechanism with
`animationFrame` in the future.

GitHub: #156