hoveredState: Rewrite to be much more efficient
Instead of doing a hitTest for each hoverable component, we now do a
single one per Window and distribut the result.

This also removes the `hitTest` argument because it is now an inherent
part of how it functions and we never disabled it anyway (and there
isn't really any reason you'd ever want to).

This also removes the `layoutSafe` argument because the the new
`UpdateFunc` API this uses is layout safe by default, with no extra
frame delay necessary.

Source-Commit: 7e67d6c184370ac5633b44c74a893daba2031d32