flowContainer: Rewrite constraints
The new backend takes an approach similar to Arrangement where the
layout of the entire container is computed once and the constraints are
merely thin accessors for that layout info. This should be much more
efficient than computing the relevant parts of the layout again and
again in each constraint independently.
This also allows for us to use Arrangement to arrange items in a row
left-to-right, as well as Alignment to align them top-to-bottom.

Source-Commit: 7e7b92871ea2094e37d7bce9cd0c1fb5972ffdb5