AlphaEffect: Remove
As of 1.21.5, the MC framebuffer is no longer bound all the time, only
during MC's draw calls. As such, the `glCopyTexSubImage2D` call will
actually copy from the wrong source.

This commit fixes the issue by using our (for now internal because still
evolving) texture api which has an explicit `mcFrameBufferColorTexture`
and as such also works properly on 1.21.5.

This unfortunately means that it can no longer be available in Elementa
unstable which because that can't have Minecraft dependencies until we
stabilize our texture API into UniversalCraft.

Source-Commit: 097d1a629a54177b19212cb81a6c79c9bae37b08